Add a dynamic importable gh app to github module
This approach allows users to access current GitHub App
instance accessible from the context var dynamically.
It can be pre-populated by the machinery once the instance
is aqcuired and initialized so that all subtasks could easily
access it.

>>> from . import github
>>> print('in handler')
>>> print(github.app)

There's a drawback, though. If users import the `app` object
directly, instead of a module, it'll be glued to a static value
rather then retrievable on-acccess.

This is why we're disregarding this approach in favor of the one
submitted in the next commit. Current commit is put into the Git
tree for historical and educational purposes, and to provide a
context wrt the decision made.